feat(server/cs-storage): refactor to GCS  service and file name formatting utility and add deleteFileFromGCS in GCS service
